+++ src/main/java/io/leavesfly/tinydl/nnet/Block.java
@@ -14,23 +14,23 @@ import java.util.Map;
  */
 public abstract class Block extends LayerAble {
 
-    protected String name;
-
-    protected Map<String, Parameter> params;
-
-    protected Shape xInputShape;
-
-    protected Shape yOutputShape;
-
     protected List<LayerAble> layers;
 
-    public Block(String _name, Shape _xInputShape, Shape _yOutputShape) {
+    public Block(String _name, Shape _inputShape) {
         name = _name;
         this.params = new HashMap<>();
         layers = new ArrayList<>();
-        xInputShape = _xInputShape;
-        yOutputShape = _yOutputShape;
+        inputShape = _inputShape;
     }
+
+    public Block(String _name, Shape _inputShape, Shape _outputShape) {
+        name = _name;
+        this.params = new HashMap<>();
+        layers = new ArrayList<>();
+        inputShape = _inputShape;
+        outputShape = _outputShape;
+    }
+
 
     @Override
     public void clearGrads() {
@@ -81,31 +81,6 @@ public abstract class Block extends LayerAble {
         }
     }
 
-    @Override
-    public Shape getInputShape() {
-        return xInputShape;
-    }
-
-    @Override
-    public Shape getOutputShape() {
-        return yOutputShape;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Map<String, Parameter> getParams() {
-        return params;
-    }
-
-    public void addParam(String paramName, Parameter value) {
-        getParams().put(name + "." + paramName, value);
-    }
-
-    public Parameter getParamBy(String paramName) {
-        return getParams().get(name + "." + paramName);
-    }
 
     public List<LayerAble> getLayers() {
         return layers;
